correct version; bugs fixed

+++ src/DataTypes.java
@@ -1,10 +1,10 @@
 import java.util.List;
 
 public class DataTypes {
-    // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
+    // to do TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
@@ -12,3 +12,5 @@ public class DataTypes {
         return s;
     }
 }
+
+// check lab 2 other file for og code + explanation on why this code works
